Party-Up
PartyUp is a JavaScript project that merges the capabilities of the React and React Native frameworks. This endeavor is the collaborative result of our team's efforts, crafted with the specific purpose of being utilized as an integral component of our Technical Capstone Project (TCC) within the context of the Technical System Development program at Etec Dr. Ruth Cardoso.